requirements.txt
fastapi
uvicorn
python-dotenv
langchain<0.3.0
langchain-community<0.3.0
langchain-core<0.3.0
langchain-google-genai
google-generativeai
rank_bm25
langchain-chroma
pydantic
requests
importlib-metadata>=6.0.0
zipp>=3.0.0
typing-extensions>=4.0.0
